# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel REST API V4 - Master-Detail Architecture",
        "description": "REST API untuk Angular Frontend dengan Master-Detail Architecture (Many-to-Many)<br><br>\n *                  <strong>Fitur V4:</strong><br>\n *                  ✅ Business Units & Menus sebagai Master Data<br>\n *                  ✅ Many-to-Many relationships (User ↔ BU, User ↔ Menu)<br>\n *                  ✅ Dropdown API untuk manage user access<br>\n *                  ✅ Dynamic menu & BU assignment<br><br>\n *                  <strong>Test Accounts:</strong><br>\n *                  - admin / Admin123 (Full access)<br>\n *                  - user1 / User123 (Batam, Jakarta + 3 menus)<br>\n *                  - user2 / User123 (Surabaya + 2 menus)",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Login user dan dapatkan token. Setelah login, user harus pilih business unit.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "User123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login berhasil. Silakan pilih business unit."
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "username": {
                                                            "type": "string",
                                                            "example": "user1"
                                                        },
                                                        "level": {
                                                            "type": "string",
                                                            "example": "user"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abc123..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Username atau password salah"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "Logout dan hapus token",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout berhasil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/business-units": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get user's accessible business units",
                "description": "Ambil daftar business unit yang boleh diakses user (V4 - many-to-many)",
                "operationId": "642065de48142289c2f5a330f0aa0ef4",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business units yang boleh diakses"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "business_unit": {
                                                        "type": "string",
                                                        "example": "Batam"
                                                    },
                                                    "active": {
                                                        "type": "string",
                                                        "example": "y"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/menus": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get user's accessible menus",
                "description": "Ambil daftar menu yang boleh diakses user (V4 - many-to-many)",
                "operationId": "a9441f9bcd7f54fd927f61557a93dcc1",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Menus yang boleh diakses"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama_menu": {
                                                        "type": "string",
                                                        "example": "Dashboard"
                                                    },
                                                    "url_link": {
                                                        "type": "string",
                                                        "example": "/dashboard"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/select-business-unit": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Select business unit",
                "description": "Pilih business unit setelah login. Token akan di-update dengan business_unit_id.",
                "operationId": "648282a1c43effcb2628f426f565d71a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "business_unit_id"
                                ],
                                "properties": {
                                    "business_unit_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business unit berhasil dipilih",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business unit berhasil dipilih: Batam"
                                        },
                                        "data": {
                                            "properties": {
                                                "business_unit": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "business_unit": {
                                                            "type": "string",
                                                            "example": "Batam"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User tidak punya akses ke business unit ini"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}